+++ iamf/cli/codec/tests/flac_encoder_test.cc
@@ -30,6 +30,7 @@ namespace {
 
 using ::absl_testing::IsOk;
 
+constexpr bool kOverrideAudioRollDistance = true;
 constexpr bool kIgnoredValidateCodecDelay = true;
 
 class FlacEncoderTest : public EncoderTestBase, public testing::Test {
@@ -48,11 +49,10 @@ class FlacEncoderTest : public EncoderTestBase, public testing::Test {
     // output are `num_samples_per_frame` and `decoder_config`.
     const CodecConfig temp = {.codec_id = CodecConfig::kCodecIdFlac,
                               .num_samples_per_frame = num_samples_per_frame_,
-                              .audio_roll_distance = 0,
                               .decoder_config = flac_decoder_config_};
 
     CodecConfigObu codec_config(ObuHeader(), 0, temp);
-    ASSERT_THAT(codec_config.Initialize(), IsOk());
+    ASSERT_THAT(codec_config.Initialize(kOverrideAudioRollDistance), IsOk());
 
     encoder_ = std::make_unique<FlacEncoder>(flac_encoder_metadata_,
                                              codec_config, num_channels_);