+++ iamf/cli/codec/tests/lpcm_decoder_test.cc
@@ -16,12 +16,13 @@ namespace {
 
 using ::absl_testing::IsOk;
 
+constexpr bool kOverrideAudioRollDistance = true;
+
 CodecConfigObu CreateCodecConfigObu(LpcmDecoderConfig lpcm_decoder_config,
                                     uint32_t num_samples_per_frame = 1024) {
   const CodecConfig codec_config = {
       .codec_id = CodecConfig::kCodecIdLpcm,
       .num_samples_per_frame = num_samples_per_frame,
-      .audio_roll_distance = 0,
       .decoder_config = lpcm_decoder_config};
 
   CodecConfigObu codec_config_obu(ObuHeader(), 0, codec_config);
@@ -35,7 +36,7 @@ TEST(LpcmDecoderTest, Construct) {
   lpcm_decoder_config.sample_format_flags_bitmask_ =
       LpcmDecoderConfig::LpcmFormatFlagsBitmask::kLpcmLittleEndian;
   CodecConfigObu codec_config_obu = CreateCodecConfigObu(lpcm_decoder_config);
-  ASSERT_THAT(codec_config_obu.Initialize(), IsOk());
+  ASSERT_THAT(codec_config_obu.Initialize(kOverrideAudioRollDistance), IsOk());
   int number_of_channels = 11;  // Arbitrary.
 
   LpcmDecoder lpcm_decoder(codec_config_obu, number_of_channels);
@@ -51,7 +52,7 @@ TEST(LpcmDecoderTest, Initialize_InvalidConfigFails) {
   lpcm_decoder_config.sample_format_flags_bitmask_ =
       LpcmDecoderConfig::LpcmFormatFlagsBitmask::kLpcmBeginReserved;
   CodecConfigObu codec_config_obu = CreateCodecConfigObu(lpcm_decoder_config);
-  ASSERT_THAT(codec_config_obu.Initialize(), IsOk());
+  ASSERT_THAT(codec_config_obu.Initialize(kOverrideAudioRollDistance), IsOk());
   int number_of_channels = 11;  // Arbitrary.
 
   LpcmDecoder lpcm_decoder(codec_config_obu, number_of_channels);
@@ -72,7 +73,7 @@ LpcmDecoder CreateDecoderForDecodingTest(uint8_t sample_size,
   lpcm_decoder_config.sample_format_flags_bitmask_ =
       little_endian ? kLpcmLittleEndian : kLpcmBigEndian;
   CodecConfigObu codec_config_obu = CreateCodecConfigObu(lpcm_decoder_config);
-  if (!codec_config_obu.Initialize().ok()) {
+  if (!codec_config_obu.Initialize(kOverrideAudioRollDistance).ok()) {
     LOG(ERROR) << "Failed to initialize codec config OBU";
   }
   int number_of_channels = 2;  // Keep the amount of test data reasonable.